(feat) Always lowercase CNAME responses
It looks like in AWS (at least, in ECS), requests for 0x20 encoded
domains under amazonaws.com that are CNAME-d, come back with a CNAME
with the same 0x20 encoded version of amazonaws.com. To ensure we use
the same cache of CNAME responses for all requests, we lowercase the
responses before processing them further.